Add enemy and health variable
Kick now connects to Apply Damage node, and the enemy BP receives this through Event AnyDamage node. Enemy is given a basic ragdoll animation upon death.

Enemy is given a basic push back when hit but not death.

The issue is the enemy pushback motion is determined by the enemy actor forward vector and enemy actor up vector, so even if you kick the enemy from the back, the enemy still moves back -300 units from the direction the enemy is facing and +200 units from the enemy up vector. So if you kick an enemy from the back and from a higher elevation for example, the enemy moves towards the kick and up instead of moving away from the kick and down.