Add basic enemy AI and attack
Enemy now chases the player character until it reaches it. Upon reaching the player, enemy character attacks the player using the kick1 attack.

Had to redo the ABP_Enemy AnimGraph because it did not have a Slot'DefaultSlot' to allow it to do an animation montage, and as a result, it was not playing the Kick1 animation even though the attack itself was successful.

"In Unreal Engine's Animation Blueprints, montages have a built-in ability to override other animations because they are played in Slots, which are designed to have higher priority than your base animations, like those from a BlendSpace. When you play an animation montage, it will automatically use the slot specified in the montage setup, and if the Animation Graph is set up with the same slot, the montage will play." -- because the animation blueprint did not have a Slot, the montage animation did not play until I added Slot'DefaultSlot' to ABP_Enemy.