Add UI elements: WB_HealthBar and WB_Hud.
WB_Health bar tracks player health using gethealth, divides the result by max health using getmaxhealth, and displays the result as a percent in the form of a progress bar in the viewport.

WB_Hud currently only contains a brief red flash when the player takes damage. The animation for this is done, but it is not connected to BP_Player events yet.